+++ CHANGELOG.md
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `APP_GOLIOTH_CONNECT_TIMEOUT_S` timeout for connecting to Golioth
+
+### Fixed
+
+- Fix a hang when waiting for settings to sync from the server
+
 ## [2.2.4] - 2025-08-29
 
 ### Fixed